move card to molecules and create css file. add on active grabbing cursor.
