+++ src/domains/datamodel/DataModelController.tsx
@@ -1,9 +1,9 @@
 import React, { FunctionComponent } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { DavitCard, DavitCardProps } from "../../components/common/fragments/DavitCard";
 import { DnDBox, DnDBoxElement, DnDBoxType } from "../../components/common/fragments/DnDBox";
 import { DavitPathProps, DavitPathTypes } from "../../components/common/fragments/svg/DavitPath";
 import { ViewPlaceholder } from "../../components/layout/ViewPlaceholder";
+import { DavitCard, DavitCardProps } from "../../components/molecules/card/DavitCard";
 import { ActorCTO } from "../../dataAccess/access/cto/ActorCTO";
 import { DataCTO } from "../../dataAccess/access/cto/DataCTO";
 import { DataSetupCTO } from "../../dataAccess/access/cto/DataSetupCTO";
